demohouse/chat2cartoon/backend/pyproject.toml
[tool.poetry]
name = "chat2cartoon"
version = "0.1.0"
description = "Fun and interactive bilingual learning"
authors = ["zhiyue.yi <zhiyue.yi@bytedance.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.0,<3.12.0"
cryptography = "^43.0.3"
tos = "2.8.1"
websockets = "13.1"
moviepy = "2.1.1"
arkitect = ">=0.0.1"
python-dotenv = "^1.0.1"
volcengine-python-sdk = "1.0.120"
volcengine = "1.0.167"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
pytest-watcher = "^0.4.3"
syrupy = "^4.7.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = "True"
ignore_missing_imports = "True"
exclude = ["tests", "scripts"]

[tool.coverage.run]
omit = [
    "tests/*",
]

[tool.pytest.ini_options]
pythonpath = "."
# --strict-markers will raise errors on unknown marks.
# https://docs.pytest.org/en/7.1.x/how-to/mark.html#raising-errors-on-unknown-marks
#
# https://docs.pytest.org/en/7.1.x/reference/reference.html
# --strict-config       any warnings encountered while parsing the `pytest`
#                       section of the configuration file raise errors.
#
# https://github.com/tophat/syrupy
# --snapshot-warn-unused    Prints a warning on unused snapshots rather than fail the test suite.
addopts = "--snapshot-warn-unused --strict-markers --strict-config --durations=5"
# Registering custom markers.
# https://docs.pytest.org/en/7.1.x/example/markers.html#registering-markers
markers = [
    "requires: mark tests as requiring a specific library",
    "asyncio: mark tests as requiring asyncio",
    "compile: mark placeholder test used to compile integration tests without running them",
]
asyncio_mode = "auto"

